pyproject.toml
[tool.poetry]
name = "wpsios"
version = "0.1.0"
description = ""
authors = ["qu3vipon <sdh5813@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3.0"
boto3 = "^1.12.23"
psycopg2-binary = "^2.8.4"
django-storages = "^1.9.1"
gunicorn = "^20.0.4"
supervisor = "^4.1.0"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"
pillow = "^7.0.0"
sentry-sdk = "^0.14.3"
drf-yasg = "^1.17.1"
coolsms_python_sdk = "^2.0.3"
django-phonenumber-field = "^4.0.0"
phonenumbers = "^8.12.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.2"
django-extensions = "^2.2.8"
httpie = "^2.0.0"
requests = "^2.23.0"
bs4 = "^0.0.1"
selenium = "^3.141.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
